@@ -426,8 +426,8 @@ table_rows: list[Row] = [
         str,
         python_input=True,
         json_input=True,
-        condition='Format: `YYYY-MM-DDTHH:MM:SS.f`. See [speedate](https://docs.rs/speedate/latest/speedate/).',
-        valid_examples=['2017-05-05 10:10:10', '2017-05-05T10:10:10.0002', '2017-05-05 10:10:10+00:00'],
+        condition='Format: `YYYY-MM-DDTHH:MM:SS.f` or `YYYY-MM-DD`. See [speedate](https://docs.rs/speedate/latest/speedate/).',
+        valid_examples=['2017-05-05 10:10:10', '2017-05-05T10:10:10.0002', '2017-05-05 10:10:10+00:00', '2017-05-05'],
         invalid_examples=['2017-5-5T10:10:10'],
         core_schemas=[core_schema.DatetimeSchema],
     ),
@@ -436,7 +436,7 @@ table_rows: list[Row] = [
         bytes,
         python_input=True,
         condition=(
-            'Format: `YYYY-MM-DDTHH:MM:SS.f`. See [speedate](https://docs.rs/speedate/latest/speedate/), (UTF-8).'
+            'Format: `YYYY-MM-DDTHH:MM:SS.f` or `YYYY-MM-DD`. See [speedate](https://docs.rs/speedate/latest/speedate/), (UTF-8).'
         ),
         valid_examples=[b'2017-05-05 10:10:10', b'2017-05-05T10:10:10.0002', b'2017-05-05 10:10:10+00:00'],
         invalid_examples=[b'2017-5-5T10:10:10'],
@@ -682,50 +682,50 @@ table_rows: list[Row] = [
         tuple,
         strict=True,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         'Array',
         strict=True,
         json_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         list,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         set,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         frozenset,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         deque,
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         'dict_keys',
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         tuple,
         'dict_values',
         python_input=True,
-        core_schemas=[core_schema.TuplePositionalSchema, core_schema.TupleVariableSchema],
+        core_schemas=[core_schema.TupleSchema],
     ),
     Row(
         set,
