@@ -17,10 +17,11 @@ if typing.TYPE_CHECKING:
 
     from . import dataclasses
     from ._internal._generate_schema import GenerateSchema as GenerateSchema
+    from .aliases import AliasChoices, AliasGenerator, AliasPath
     from .annotated_handlers import GetCoreSchemaHandler, GetJsonSchemaHandler
-    from .config import ConfigDict
+    from .config import ConfigDict, with_config
     from .errors import *
-    from .fields import AliasChoices, AliasPath, Field, PrivateAttr, computed_field
+    from .fields import Field, PrivateAttr, computed_field
     from .functional_serializers import (
         PlainSerializer,
         SerializeAsAny,
@@ -44,7 +45,7 @@ if typing.TYPE_CHECKING:
     from .type_adapter import TypeAdapter
     from .types import *
     from .validate_call_decorator import validate_call
-    from .warnings import PydanticDeprecatedSince20, PydanticDeprecationWarning
+    from .warnings import PydanticDeprecatedSince20, PydanticDeprecatedSince26, PydanticDeprecationWarning
 
     # this encourages pycharm to import `ValidationError` from here, not pydantic_core
     ValidationError = pydantic_core.ValidationError
@@ -79,6 +80,7 @@ __all__ = (
     'WrapSerializer',
     # config
     'ConfigDict',
+    'with_config',
     # deprecated V1 config, these are imported via `__getattr__` below
     'BaseConfig',
     'Extra',
@@ -92,11 +94,13 @@ __all__ = (
     'PydanticUndefinedAnnotation',
     'PydanticInvalidForJsonSchema',
     # fields
-    'AliasPath',
-    'AliasChoices',
     'Field',
     'computed_field',
     'PrivateAttr',
+    # alias
+    'AliasChoices',
+    'AliasGenerator',
+    'AliasPath',
     # main
     'BaseModel',
     'create_model',
@@ -158,6 +162,7 @@ __all__ = (
     'DirectoryPath',
     'NewPath',
     'Json',
+    'Secret',
     'SecretStr',
     'SecretBytes',
     'StrictBool',
@@ -193,6 +198,7 @@ __all__ = (
     'VERSION',
     # warnings
     'PydanticDeprecatedSince20',
+    'PydanticDeprecatedSince26',
     'PydanticDeprecationWarning',
     # annotated handlers
     'GetCoreSchemaHandler',
@@ -206,6 +212,7 @@ __all__ = (
     'ValidatorFunctionWrapHandler',
     'FieldSerializationInfo',
     'SerializerFunctionWrapHandler',
+    'OnErrorOmit',
 )
 
 # A mapping of {<member name>: (package, <module name>)} defining dynamic imports
@@ -230,6 +237,7 @@ _dynamic_imports: 'dict[str, tuple[str, str]]' = {
     'WrapSerializer': (__package__, '.functional_serializers'),
     # config
     'ConfigDict': (__package__, '.config'),
+    'with_config': (__package__, '.config'),
     # validate call
     'validate_call': (__package__, '.validate_call_decorator'),
     # errors
@@ -240,11 +248,13 @@ _dynamic_imports: 'dict[str, tuple[str, str]]' = {
     'PydanticUndefinedAnnotation': (__package__, '.errors'),
     'PydanticInvalidForJsonSchema': (__package__, '.errors'),
     # fields
-    'AliasPath': (__package__, '.fields'),
-    'AliasChoices': (__package__, '.fields'),
     'Field': (__package__, '.fields'),
     'computed_field': (__package__, '.fields'),
     'PrivateAttr': (__package__, '.fields'),
+    # alias
+    'AliasChoices': (__package__, '.aliases'),
+    'AliasGenerator': (__package__, '.aliases'),
+    'AliasPath': (__package__, '.aliases'),
     # main
     'BaseModel': (__package__, '.main'),
     'create_model': (__package__, '.main'),
@@ -302,6 +312,7 @@ _dynamic_imports: 'dict[str, tuple[str, str]]' = {
     'DirectoryPath': (__package__, '.types'),
     'NewPath': (__package__, '.types'),
     'Json': (__package__, '.types'),
+    'Secret': (__package__, '.types'),
     'SecretStr': (__package__, '.types'),
     'SecretBytes': (__package__, '.types'),
     'StrictBool': (__package__, '.types'),
@@ -330,10 +341,12 @@ _dynamic_imports: 'dict[str, tuple[str, str]]' = {
     'Tag': (__package__, '.types'),
     'Discriminator': (__package__, '.types'),
     'JsonValue': (__package__, '.types'),
+    'OnErrorOmit': (__package__, '.types'),
     # type_adapter
     'TypeAdapter': (__package__, '.type_adapter'),
     # warnings
     'PydanticDeprecatedSince20': (__package__, '.warnings'),
+    'PydanticDeprecatedSince26': (__package__, '.warnings'),
     'PydanticDeprecationWarning': (__package__, '.warnings'),
     # annotated handlers
     'GetCoreSchemaHandler': (__package__, '.annotated_handlers'),
