@@ -2,16 +2,13 @@
 from __future__ import annotations as _annotations
 
 import dataclasses
-import inspect
 import typing
 import warnings
 from functools import partial, wraps
-from inspect import Parameter, Signature
 from typing import Any, Callable, ClassVar
 
 from pydantic_core import (
     ArgsKwargs,
-    PydanticUndefined,
     SchemaSerializer,
     SchemaValidator,
     core_schema,
@@ -23,13 +20,12 @@ from ..fields import FieldInfo
 from ..plugin._schema_validator import create_schema_validator
 from ..warnings import PydanticDeprecatedSince20
 from . import _config, _decorators, _typing_extra
-from ._config import ConfigWrapper
 from ._fields import collect_dataclass_fields
-from ._generate_schema import GenerateSchema, generate_pydantic_signature
+from ._generate_schema import GenerateSchema
 from ._generics import get_standard_typevars_map
 from ._mock_val_ser import set_dataclass_mocks
 from ._schema_generation_shared import CallbackGetCoreSchemaHandler
-from ._utils import is_valid_identifier
+from ._signature import generate_pydantic_signature
 
 if typing.TYPE_CHECKING:
     from ..config import ConfigDict
@@ -69,15 +65,20 @@ else:
     DeprecationWarning = PydanticDeprecatedSince20
 
 
-def set_dataclass_fields(cls: type[StandardDataclass], types_namespace: dict[str, Any] | None = None) -> None:
+def set_dataclass_fields(
+    cls: type[StandardDataclass],
+    types_namespace: dict[str, Any] | None = None,
+    config_wrapper: _config.ConfigWrapper | None = None,
+) -> None:
     """Collect and set `cls.__pydantic_fields__`.
 
     Args:
         cls: The class.
         types_namespace: The types namespace, defaults to `None`.
+        config_wrapper: The config wrapper instance, defaults to `None`.
     """
     typevars_map = get_standard_typevars_map(cls)
-    fields = collect_dataclass_fields(cls, types_namespace, typevars_map=typevars_map)
+    fields = collect_dataclass_fields(cls, types_namespace, typevars_map=typevars_map, config_wrapper=config_wrapper)
 
     cls.__pydantic_fields__ = fields  # type: ignore
 
@@ -115,7 +116,7 @@ def complete_dataclass(
     if types_namespace is None:
         types_namespace = _typing_extra.get_cls_types_namespace(cls)
 
-    set_dataclass_fields(cls, types_namespace)
+    set_dataclass_fields(cls, types_namespace, config_wrapper=config_wrapper)
 
     typevars_map = get_standard_typevars_map(cls)
     gen_schema = GenerateSchema(
@@ -125,7 +126,12 @@ def complete_dataclass(
     )
 
     # This needs to be called before we change the __init__
-    sig = generate_dataclass_signature(cls, cls.__pydantic_fields__, config_wrapper)  # type: ignore
+    sig = generate_pydantic_signature(
+        init=cls.__init__,
+        fields=cls.__pydantic_fields__,  # type: ignore
+        config_wrapper=config_wrapper,
+        is_dataclass=True,
+    )
 
     # dataclass.__init__ must be defined here so its `__qualname__` can be changed since functions can't be copied.
     def __init__(__dataclass_self__: PydanticDataclass, *args: Any, **kwargs: Any) -> None:
@@ -185,63 +191,6 @@ def complete_dataclass(
         cls.__setattr__ = validated_setattr.__get__(None, cls)  # type: ignore
 
     return True
-
-
-def process_param_defaults(param: Parameter) -> Parameter:
-    """Custom processing where the parameter default is of type FieldInfo
-
-    Args:
-        param (Parameter): The parameter
-
-    Returns:
-        Parameter: The custom processed parameter
-    """
-    param_default = param.default
-    if isinstance(param_default, FieldInfo):
-        annotation = param.annotation
-        # Replace the annotation if appropriate
-        # inspect does "clever" things to show annotations as strings because we have
-        # `from __future__ import annotations` in main, we don't want that
-        if annotation == 'Any':
-            annotation = Any
-
-        # Replace the field name with the alias if present
-        name = param.name
-        alias = param_default.alias
-        validation_alias = param_default.validation_alias
-        if validation_alias is None and isinstance(alias, str) and is_valid_identifier(alias):
-            name = alias
-        elif isinstance(validation_alias, str) and is_valid_identifier(validation_alias):
-            name = validation_alias
-
-        # Replace the field default
-        default = param_default.default
-        if default is PydanticUndefined:
-            if param_default.default_factory is PydanticUndefined:
-                default = inspect.Signature.empty
-            else:
-                # this is used by dataclasses to indicate a factory exists:
-                default = dataclasses._HAS_DEFAULT_FACTORY  # type: ignore
-        return param.replace(annotation=annotation, name=name, default=default)
-    return param
-
-
-def generate_dataclass_signature(
-    cls: type[StandardDataclass], fields: dict[str, FieldInfo], config_wrapper: ConfigWrapper
-) -> Signature:
-    """Generate signature for a pydantic dataclass.
-
-    Args:
-        cls: The dataclass.
-        fields: The model fields.
-        config_wrapper: The config wrapper instance.
-
-    Returns:
-        The dataclass signature.
-    """
-    return generate_pydantic_signature(
-        init=cls.__init__, fields=fields, config_wrapper=config_wrapper, post_process_parameter=process_param_defaults
-    )
 
 
 def is_builtin_dataclass(_cls: type[Any]) -> TypeGuard[type[StandardDataclass]]:
