@@ -1,3 +1,4 @@
+import functools
 import importlib.util
 import re
 import sys
@@ -302,7 +303,6 @@ def test_tuple_more():
         (dict, frozenset, list, set, tuple, type),
     ],
 )
-@pytest.mark.skipif(sys.version_info < (3, 9), reason='PEP585 generics only supported for python 3.9 and above')
 def test_pep585_generic_types(dict_cls, frozenset_cls, list_cls, set_cls, tuple_cls, type_cls):
     class Type1:
         pass
@@ -312,19 +312,19 @@ def test_pep585_generic_types(dict_cls, frozenset_cls, list_cls, set_cls, tuple_
 
     class Model(BaseModel, arbitrary_types_allowed=True):
         a: dict_cls
-        a1: dict_cls[str, int]
+        a1: 'dict_cls[str, int]'
         b: frozenset_cls
-        b1: frozenset_cls[int]
+        b1: 'frozenset_cls[int]'
         c: list_cls
-        c1: list_cls[int]
+        c1: 'list_cls[int]'
         d: set_cls
-        d1: set_cls[int]
+        d1: 'set_cls[int]'
         e: tuple_cls
-        e1: tuple_cls[int]
-        e2: tuple_cls[int, ...]
-        e3: tuple_cls[()]
+        e1: 'tuple_cls[int]'
+        e2: 'tuple_cls[int, ...]'
+        e3: 'tuple_cls[()]'
         f: type_cls
-        f1: type_cls[Type1]
+        f1: 'type_cls[Type1]'
 
     default_model_kwargs = dict(
         a={},
@@ -360,7 +360,7 @@ def test_pep585_generic_types(dict_cls, frozenset_cls, list_cls, set_cls, tuple_
     assert m.f1 == Type1
 
     with pytest.raises(ValidationError) as exc_info:
-        Model(**(default_model_kwargs | {'e3': (1,)}))
+        Model(**{**default_model_kwargs, 'e3': (1,)})
     # insert_assert(exc_info.value.errors(include_url=False))
     assert exc_info.value.errors(include_url=False) == [
         {
@@ -372,10 +372,10 @@ def test_pep585_generic_types(dict_cls, frozenset_cls, list_cls, set_cls, tuple_
         }
     ]
 
-    Model(**(default_model_kwargs | {'f': Type2}))
+    Model(**{**default_model_kwargs, 'f': Type2})
 
     with pytest.raises(ValidationError) as exc_info:
-        Model(**(default_model_kwargs | {'f1': Type2}))
+        Model(**{**default_model_kwargs, 'f1': Type2})
     # insert_assert(exc_info.value.errors(include_url=False))
     assert exc_info.value.errors(include_url=False) == [
         {
@@ -1296,6 +1296,14 @@ def test_self():
     }
 
 
+def test_no_name_conflict_in_constructor():
+    class Model(BaseModel):
+        self: int
+
+    m = Model(**{'__pydantic_self__': 4, 'self': 2})
+    assert m.self == 2
+
+
 def test_self_recursive():
     class SubModel(BaseModel):
         self: int
@@ -1366,19 +1374,17 @@ def test_type_on_annotation():
         pass
 
     class Model(BaseModel):
-        a: int = int
-        b: Type[int]
-        c: Type[int] = int
-        d: FooBar = FooBar
-        e: Type[FooBar]
-        f: Type[FooBar] = FooBar
-        g: Sequence[Type[FooBar]] = [FooBar]
-        h: Union[Type[FooBar], Sequence[Type[FooBar]]] = FooBar
-        i: Union[Type[FooBar], Sequence[Type[FooBar]]] = [FooBar]
+        a: Type[int]
+        b: Type[int] = int
+        c: Type[FooBar]
+        d: Type[FooBar] = FooBar
+        e: Sequence[Type[FooBar]] = [FooBar]
+        f: Union[Type[FooBar], Sequence[Type[FooBar]]] = FooBar
+        g: Union[Type[FooBar], Sequence[Type[FooBar]]] = [FooBar]
 
         model_config = dict(arbitrary_types_allowed=True)
 
-    assert Model.model_fields.keys() == set('abcdefghi')
+    assert Model.model_fields.keys() == set('abcdefg')
 
 
 def test_assign_type():
@@ -2373,10 +2379,9 @@ def test_abstractmethod_missing_for_all_decorators(bases):
         Square(side=1.0)
 
 
-@pytest.mark.skipif(sys.version_info < (3, 9), reason='cannot use list.__class_getitem__ before 3.9')
 def test_generic_wrapped_forwardref():
     class Operation(BaseModel):
-        callbacks: list['PathItem']
+        callbacks: 'list[PathItem]'
 
     class PathItem(BaseModel):
         pass
@@ -2427,15 +2432,11 @@ def test_invalid_forward_ref_model():
     """
     # The problem:
     if sys.version_info >= (3, 11):
-        error = RecursionError
-        kwargs = {}
+        # See PR #8243, this was a RecursionError raised by Python, but is now caught on the Pydantic side
+        error = errors.PydanticUserError
     else:
         error = TypeError
-        kwargs = {
-            'match': r'Forward references must evaluate to types\.'
-            r' Got FieldInfo\(annotation=NoneType, required=False\)\.'
-        }
-    with pytest.raises(error, **kwargs):
+    with pytest.raises(error):
 
         class M(BaseModel):
             B: ForwardRef('B') = Field(default=None)
@@ -2474,7 +2475,6 @@ def test_invalid_forward_ref_model():
     ]
 
 
-@pytest.mark.skipif(sys.version_info < (3, 9), reason='cannot parametrize types before 3.9')
 @pytest.mark.parametrize(
     ('sequence_type', 'input_data', 'expected_error_type', 'expected_error_msg', 'expected_error_ctx'),
     [
@@ -2538,8 +2538,8 @@ def test_multiple_enums():
         (Literal[False], str, False, 'false', False, 'false'),
         (Literal[True], str, True, 'true', True, 'true'),
         (Literal[False], str, 'abc', '"abc"', 'abc', '"abc"'),
-        (Literal[False], int, False, 'false', 0, '0'),
-        (Literal[True], int, True, 'true', 1, '1'),
+        (Literal[False], int, False, 'false', False, 'false'),
+        (Literal[True], int, True, 'true', True, 'true'),
         (Literal[False], int, 42, '42', 42, '42'),
     ],
 )
@@ -2704,3 +2704,31 @@ def test_recursive_root_models_in_discriminated_union():
         'title': 'Outer',
         'type': 'object',
     }
+
+
+def test_eq_with_cached_property():
+    """
+    Test BaseModel.__eq__ compatibility with functools.cached_property
+
+    See GH-7444: https://github.com/pydantic/pydantic/issues/7444
+    Previously, pydantic BaseModel.__eq__ compared the full __dict__ of
+    model instances. This is not compatible with e.g. functools.cached_property,
+    which caches the computed values in the instance's __dict__
+    """
+
+    class Model(BaseModel):
+        attr: int
+
+        @functools.cached_property
+        def cached(self) -> int:
+            return 0
+
+    obj1 = Model(attr=1)
+    obj2 = Model(attr=1)
+    # ensure the instances are indeed equal before __dict__ mutations
+    assert obj1 == obj2
+    # This access to the cached_property has the side-effect of modifying obj1.__dict__
+    # See functools.cached_property documentation and source code
+    obj1.cached
+    # Ensure the objects still compare equals after caching a property
+    assert obj1 == obj2
