@@ -38,9 +38,9 @@ from uuid import UUID
 import annotated_types
 import dirty_equals
 import pytest
-from dirty_equals import HasRepr, IsOneOf, IsStr
+from dirty_equals import HasRepr, IsFloatNan, IsOneOf, IsStr
 from pydantic_core import CoreSchema, PydanticCustomError, SchemaError, core_schema
-from typing_extensions import Annotated, Literal, TypedDict, get_args
+from typing_extensions import Annotated, Literal, NotRequired, TypedDict, get_args
 
 from pydantic import (
     UUID1,
@@ -48,6 +48,7 @@ from pydantic import (
     UUID4,
     UUID5,
     AfterValidator,
+    AllowInfNan,
     AwareDatetime,
     Base64Bytes,
     Base64Str,
@@ -64,6 +65,8 @@ from pydantic import (
     FutureDate,
     FutureDatetime,
     GetCoreSchemaHandler,
+    GetPydanticSchema,
+    ImportString,
     InstanceOf,
     Json,
     JsonValue,
@@ -76,20 +79,26 @@ from pydantic import (
     NonNegativeInt,
     NonPositiveFloat,
     NonPositiveInt,
+    OnErrorOmit,
     PastDate,
     PastDatetime,
+    PlainSerializer,
     PositiveFloat,
     PositiveInt,
     PydanticInvalidForJsonSchema,
+    PydanticSchemaGenerationError,
+    Secret,
     SecretBytes,
     SecretStr,
     SerializeAsAny,
     SkipValidation,
+    Strict,
     StrictBool,
     StrictBytes,
     StrictFloat,
     StrictInt,
     StrictStr,
+    StringConstraints,
     Tag,
     TypeAdapter,
     UTCDatetime,
@@ -108,9 +117,6 @@ from pydantic import (
     validate_call,
 )
 from pydantic.dataclasses import dataclass as pydantic_dataclass
-from pydantic.errors import PydanticSchemaGenerationError
-from pydantic.functional_validators import AfterValidator
-from pydantic.types import AllowInfNan, GetPydanticSchema, ImportString, Strict, StringConstraints
 
 try:
     import email_validator
@@ -874,6 +880,8 @@ def test_string_import_callable(annotation):
     [
         ('math:cos', 'math.cos', 'json'),
         ('math:cos', math.cos, 'python'),
+        ('math.cos', 'math.cos', 'json'),
+        ('math.cos', math.cos, 'python'),
         pytest.param(
             'os.path', 'posixpath', 'json', marks=pytest.mark.skipif(sys.platform == 'win32', reason='different output')
         ),
@@ -898,6 +906,22 @@ def test_string_import_any(value: Any, expected: Any, mode: Literal['json', 'pyt
         thing: ImportString
 
     assert PyObjectModel(thing=value).model_dump(mode=mode) == {'thing': expected}
+
+
+@pytest.mark.parametrize(
+    ('value', 'validate_default', 'expected'),
+    [
+        (math.cos, True, math.cos),
+        ('math:cos', True, math.cos),
+        (math.cos, False, math.cos),
+        ('math:cos', False, 'math:cos'),
+    ],
+)
+def test_string_import_default_value(value: Any, validate_default: bool, expected: Any):
+    class PyObjectModel(BaseModel):
+        thing: ImportString = Field(default=value, validate_default=validate_default)
+
+    assert PyObjectModel().thing == expected
 
 
 @pytest.mark.parametrize('value', ['oss', 'os.os', f'{__name__}.x'])
@@ -1457,9 +1481,9 @@ def test_datetime_errors(DatetimeModel):
     # insert_assert(exc_info.value.errors(include_url=False))
     assert exc_info.value.errors(include_url=False) == [
         {
-            'type': 'datetime_parsing',
+            'type': 'datetime_from_date_parsing',
             'loc': ('dt',),
-            'msg': 'Input should be a valid datetime, month value is outside expected range of 1-12',
+            'msg': 'Input should be a valid datetime or date, month value is outside expected range of 1-12',
             'input': '2017-13-05T19:47:07',
             'ctx': {'error': 'month value is outside expected range of 1-12'},
         },
@@ -2368,6 +2392,10 @@ def test_sequence_fails(cls, value, errors):
     assert exc_info.value.errors(include_url=False) == errors
 
 
+def test_sequence_strict():
+    assert TypeAdapter(Sequence[int]).validate_python((), strict=True) == ()
+
+
 def test_int_validation():
     class Model(BaseModel):
         a: PositiveInt = None
@@ -2530,13 +2558,34 @@ def test_float_validation():
     ]
 
 
-def test_finite_float_validation():
+def test_infinite_float_validation():
     class Model(BaseModel):
         a: float = None
 
     assert Model(a=float('inf')).a == float('inf')
     assert Model(a=float('-inf')).a == float('-inf')
     assert math.isnan(Model(a=float('nan')).a)
+
+
+@pytest.mark.parametrize(
+    ('ser_json_inf_nan', 'input', 'output', 'python_roundtrip'),
+    (
+        ('null', float('inf'), 'null', None),
+        ('null', float('-inf'), 'null', None),
+        ('null', float('nan'), 'null', None),
+        ('constants', float('inf'), 'Infinity', float('inf')),
+        ('constants', float('-inf'), '-Infinity', float('-inf')),
+        ('constants', float('nan'), 'NaN', IsFloatNan),
+    ),
+)
+def test_infinite_float_json_serialization(ser_json_inf_nan, input, output, python_roundtrip):
+    class Model(BaseModel):
+        model_config = ConfigDict(ser_json_inf_nan=ser_json_inf_nan)
+        a: float
+
+    json_string = Model(a=input).model_dump_json()
+    assert json_string == f'{{"a":{output}}}'
+    assert json.loads(json_string) == {'a': python_roundtrip}
 
 
 @pytest.mark.parametrize('value', [float('inf'), float('-inf'), float('nan')])
@@ -4053,6 +4102,166 @@ def test_secretstr_idempotent():
     assert m.password.get_secret_value() == '1234'
 
 
+class SecretDate(Secret[date]):
+    def _display(self) -> str:
+        return '****/**/**'
+
+
+class SampleEnum(str, Enum):
+    foo = 'foo'
+    bar = 'bar'
+
+
+SecretEnum = Secret[SampleEnum]
+
+
+@pytest.mark.parametrize(
+    'value, result',
+    [
+        # Valid inputs
+        (1_493_942_400, date(2017, 5, 5)),
+        (1_493_942_400_000, date(2017, 5, 5)),
+        (0, date(1970, 1, 1)),
+        ('2012-04-23', date(2012, 4, 23)),
+        (b'2012-04-23', date(2012, 4, 23)),
+        (date(2012, 4, 9), date(2012, 4, 9)),
+        (datetime(2012, 4, 9, 0, 0), date(2012, 4, 9)),
+        (1_549_238_400, date(2019, 2, 4)),  # nowish in s
+        (1_549_238_400_000, date(2019, 2, 4)),  # nowish in ms
+        (19_999_958_400, date(2603, 10, 11)),  # just before watershed
+    ],
+)
+def test_secretdate(value, result):
+    class Foobar(BaseModel):
+        value: SecretDate
+
+    f = Foobar(value=value)
+
+    # Assert correct type.
+    assert f.value.__class__.__name__ == 'SecretDate'
+
+    # Assert str and repr are correct.
+    assert str(f.value) == '****/**/**'
+    assert repr(f.value) == "SecretDate('****/**/**')"
+
+    # Assert retrieval of secret value is correct
+    assert f.value.get_secret_value() == result
+
+
+def test_secretdate_json_serializable():
+    class _SecretDate(Secret[date]):
+        def _display(self) -> str:
+            return '****/**/**'
+
+    SecretDate = Annotated[
+        _SecretDate,
+        PlainSerializer(lambda v: v.get_secret_value().strftime('%Y-%m-%d'), when_used='json'),
+    ]
+
+    class Foobar(BaseModel):
+        value: SecretDate
+
+    f = Foobar(value='2017-01-01')
+
+    assert '2017-01-01' in f.model_dump_json()
+
+
+def test_secretenum_json_serializable():
+    class SampleEnum(str, Enum):
+        foo = 'foo'
+        bar = 'bar'
+
+    SecretEnum = Annotated[
+        Secret[SampleEnum],
+        PlainSerializer(lambda v: v.get_secret_value(), when_used='json'),
+    ]
+
+    class Foobar(BaseModel):
+        value: SecretEnum
+
+    f = Foobar(value='foo')
+
+    assert f.model_dump_json() == '{"value":"foo"}'
+
+
+@pytest.mark.parametrize(
+    'SecretField, value, error_msg',
+    [
+        (SecretDate, 'not-a-date', r'Input should be a valid date'),
+        (SecretStr, 0, r'Input should be a valid string \[type=string_type,'),
+        (SecretBytes, 0, r'Input should be a valid bytes \[type=bytes_type,'),
+        (SecretEnum, 0, r'Input should be an instance of SampleEnum'),
+    ],
+)
+def test_strict_secretfield_by_config(SecretField, value, error_msg):
+    class Foobar(BaseModel):
+        model_config = ConfigDict(strict=True)
+        value: SecretField
+
+    with pytest.raises(ValidationError, match=error_msg):
+        Foobar(value=value)
+
+
+@pytest.mark.parametrize(
+    'field, value, error_msg',
+    [
+        (date, 'not-a-date', r'Input should be a valid date'),
+        (str, 0, r'Input should be a valid string \[type=string_type,'),
+        (bytes, 0, r'Input should be a valid bytes \[type=bytes_type,'),
+        (SampleEnum, 0, r'Input should be an instance of SampleEnum'),
+    ],
+)
+def test_strict_secretfield_annotated(field, value, error_msg):
+    SecretField = Annotated[field, Strict()]
+
+    class Foobar(BaseModel):
+        value: Secret[SecretField]
+
+    with pytest.raises(ValidationError, match=error_msg):
+        Foobar(value=value)
+
+
+@pytest.mark.parametrize(
+    'value',
+    [
+        datetime(2012, 4, 9, 12, 15),
+        'x20120423',
+        '2012-04-56',
+        20000044800,  # just after watershed
+        1_549_238_400_000_000,  # nowish in μs
+        1_549_238_400_000_000_000,  # nowish in ns
+        'infinity',
+        float('inf'),
+        int('1' + '0' * 100),
+        1e1000,
+        float('-infinity'),
+        float('nan'),
+    ],
+)
+def test_secretdate_parsing(value):
+    class FooBar(BaseModel):
+        d: SecretDate
+
+    with pytest.raises(ValidationError):
+        FooBar(d=value)
+
+
+def test_secretdate_equality():
+    assert SecretDate('2017-01-01') == SecretDate('2017-01-01')
+    assert SecretDate('2017-01-01') != SecretDate('2018-01-01')
+    assert SecretDate(date(2017, 1, 1)) != date(2017, 1, 1)
+    assert SecretDate('2017-01-01') is not SecretDate('2017-01-01')
+
+
+def test_secretdate_idempotent():
+    class Foobar(BaseModel):
+        value: SecretDate
+
+    # Should not raise an exception
+    m = Foobar(value=SecretDate(date(2017, 1, 1)))
+    assert m.value.get_secret_value() == date(2017, 1, 1)
+
+
 @pytest.mark.parametrize(
     'pydantic_type',
     [
@@ -4072,6 +4281,7 @@ def test_secretstr_idempotent():
         StrictFloat,
         FiniteFloat,
         conbytes,
+        Secret,
         SecretBytes,
         constr,
         StrictStr,
@@ -4170,6 +4380,13 @@ def test_secret_str_min_max_length():
 
     value = '1' * 8
     assert Foobar(password=value).password.get_secret_value() == value
+
+
+def test_secretbytes_json():
+    class Foobar(BaseModel):
+        password: SecretBytes
+
+    assert Foobar(password='foo').model_dump_json() == '{"password":"**********"}'
 
 
 def test_secretbytes():
@@ -4391,20 +4608,23 @@ def test_frozenset_field_not_convertible():
 
 
 @pytest.mark.parametrize(
-    'input_value,output,human_bin,human_dec',
+    'input_value,output,human_bin,human_dec,human_sep',
     (
-        ('1', 1, '1B', '1B'),
-        ('1.0', 1, '1B', '1B'),
-        ('1b', 1, '1B', '1B'),
-        ('1.5 KB', int(1.5e3), '1.5KiB', '1.5KB'),
-        ('1.5 K', int(1.5e3), '1.5KiB', '1.5KB'),
-        ('1.5 MB', int(1.5e6), '1.4MiB', '1.5MB'),
-        ('1.5 M', int(1.5e6), '1.4MiB', '1.5MB'),
-        ('5.1kib', 5222, '5.1KiB', '5.2KB'),
-        ('6.2EiB', 7148113328562451456, '6.2EiB', '7.1EB'),
+        (1, 1, '1B', '1B', '1 B'),
+        ('1', 1, '1B', '1B', '1 B'),
+        ('1.0', 1, '1B', '1B', '1 B'),
+        ('1b', 1, '1B', '1B', '1 B'),
+        ('1.5 KB', int(1.5e3), '1.5KiB', '1.5KB', '1.5 KiB'),
+        ('1.5 K', int(1.5e3), '1.5KiB', '1.5KB', '1.5 KiB'),
+        ('1.5 MB', int(1.5e6), '1.4MiB', '1.5MB', '1.4 MiB'),
+        ('1.5 M', int(1.5e6), '1.4MiB', '1.5MB', '1.4 MiB'),
+        ('5.1kib', 5222, '5.1KiB', '5.2KB', '5.1 KiB'),
+        ('6.2EiB', 7148113328562451456, '6.2EiB', '7.1EB', '6.2 EiB'),
+        ('8bit', 1, '1B', '1B', '1 B'),
+        ('1kbit', 125, '125B', '125B', '125 B'),
     ),
 )
-def test_bytesize_conversions(input_value, output, human_bin, human_dec):
+def test_bytesize_conversions(input_value, output, human_bin, human_dec, human_sep):
     class Model(BaseModel):
         size: ByteSize
 
@@ -4414,6 +4634,7 @@ def test_bytesize_conversions(input_value, output, human_bin, human_dec):
 
     assert m.size.human_readable() == human_bin
     assert m.size.human_readable(decimal=True) == human_dec
+    assert m.size.human_readable(separator=' ') == human_sep
 
 
 def test_bytesize_to():
@@ -4425,17 +4646,36 @@ def test_bytesize_to():
     assert m.size.to('MiB') == pytest.approx(1024)
     assert m.size.to('MB') == pytest.approx(1073.741824)
     assert m.size.to('TiB') == pytest.approx(0.0009765625)
+    assert m.size.to('bit') == pytest.approx(8589934592)
+    assert m.size.to('kbit') == pytest.approx(8589934.592)
 
 
 def test_bytesize_raises():
     class Model(BaseModel):
         size: ByteSize
 
-    with pytest.raises(ValidationError, match='parse value'):
+    with pytest.raises(ValidationError, match='parse value') as exc_info:
         Model(size='d1MB')
+    assert exc_info.value.errors(include_url=False) == [
+        {
+            'input': 'd1MB',
+            'loc': ('size',),
+            'msg': 'could not parse value and unit from byte string',
+            'type': 'byte_size',
+        }
+    ]
 
-    with pytest.raises(ValidationError, match='byte unit'):
+    with pytest.raises(ValidationError, match='byte unit') as exc_info:
         Model(size='1LiB')
+    assert exc_info.value.errors(include_url=False) == [
+        {
+            'ctx': {'unit': 'LiB'},
+            'input': '1LiB',
+            'loc': ('size',),
+            'msg': 'could not interpret byte unit: LiB',
+            'type': 'byte_size_unit',
+        }
+    ]
 
     # 1Gi is not a valid unit unlike 1G
     with pytest.raises(ValidationError, match='byte unit'):
@@ -4444,6 +4684,9 @@ def test_bytesize_raises():
     m = Model(size='1MB')
     with pytest.raises(PydanticCustomError, match='byte unit'):
         m.size.to('bad_unit')
+
+    with pytest.raises(PydanticCustomError, match='byte unit'):
+        m.size.to('1ZiB')
 
 
 def test_deque_success():
@@ -5633,7 +5876,8 @@ def test_typing_literal_field():
 
 def test_instance_of_annotation():
     class Model(BaseModel):
-        x: InstanceOf[Sequence[int]]  # Note: the generic parameter gets ignored by runtime validation
+        # Note: the generic parameter gets ignored by runtime validation
+        x: InstanceOf[Sequence[int]]
 
     class MyList(list):
         pass
@@ -6043,7 +6287,8 @@ def test_union_tags_in_errors():
     with pytest.raises(ValidationError) as exc_info:
         adapter.validate_python(['a'])
 
-    assert '2 validation errors for union[function-after[<lambda>(), list[int]],dict[str,str]]' in str(exc_info)  # yuck
+    # yuck
+    assert '2 validation errors for union[function-after[<lambda>(), list[int]],dict[str,str]]' in str(exc_info)
     # the loc's are bad here:
     assert exc_info.value.errors(include_url=False) == [
         {
@@ -6087,7 +6332,8 @@ def test_union_tags_in_errors():
 def test_json_value():
     adapter = TypeAdapter(JsonValue)
     valid_json_data = {'a': {'b': {'c': 1, 'd': [2, None]}}}
-    invalid_json_data = {'a': {'b': ...}}  # would pass validation as a dict[str, Any]
+    # would pass validation as a dict[str, Any]
+    invalid_json_data = {'a': {'b': ...}}
 
     assert adapter.validate_python(valid_json_data) == valid_json_data
     assert adapter.validate_json(json.dumps(valid_json_data)) == valid_json_data
@@ -6102,3 +6348,68 @@ def test_json_value():
             'type': 'invalid-json-value',
         }
     ]
+
+
+def test_json_value_with_subclassed_types():
+    class IntType(int):
+        pass
+
+    class FloatType(float):
+        pass
+
+    class StrType(str):
+        pass
+
+    class ListType(list):
+        pass
+
+    class DictType(dict):
+        pass
+
+    adapter = TypeAdapter(JsonValue)
+    valid_json_data = {'int': IntType(), 'float': FloatType(), 'str': StrType(), 'list': ListType(), 'dict': DictType()}
+    assert adapter.validate_python(valid_json_data) == valid_json_data
+
+
+def test_json_value_roundtrip() -> None:
+    # see https://github.com/pydantic/pydantic/issues/8175
+    class MyModel(BaseModel):
+        val: Union[str, JsonValue]
+
+    round_trip_value = json.loads(MyModel(val=True).model_dump_json())['val']
+    assert round_trip_value is True, round_trip_value
+
+
+def test_on_error_omit() -> None:
+    OmittableInt = OnErrorOmit[int]
+
+    class MyTypedDict(TypedDict):
+        a: NotRequired[OmittableInt]
+        b: NotRequired[OmittableInt]
+
+    class Model(BaseModel):
+        a_list: List[OmittableInt]
+        a_dict: Dict[OmittableInt, OmittableInt]
+        a_typed_dict: MyTypedDict
+
+    actual = Model(
+        a_list=[1, 2, 'a', 3],
+        a_dict={1: 1, 2: 2, 'a': 'a', 'b': 0, 3: 'c', 4: 4},
+        a_typed_dict=MyTypedDict(a=1, b='xyz'),  # type: ignore
+    )
+
+    expected = Model(a_list=[1, 2, 3], a_dict={1: 1, 2: 2, 4: 4}, a_typed_dict=MyTypedDict(a=1))
+
+    assert actual == expected
+
+
+def test_on_error_omit_top_level() -> None:
+    ta = TypeAdapter(OnErrorOmit[int])
+
+    assert ta.validate_python(1) == 1
+    assert ta.validate_python('1') == 1
+
+    # we might want to just raise the OmitError or convert it to a ValidationError
+    # if it hits the top level, but this documents the current behavior at least
+    with pytest.raises(SchemaError, match='Uncaught Omit error'):
+        ta.validate_python('a')
